\name{groupinfo_ext}
\alias{groupinfo_ext}
\docType{data}
\title{
An out put of create_group
}
\description{
An output of create_group()
}
\usage{data("groupinfo_ext")}
\format{
  A data frame with 28770 observations on the following variable.
  \describe{
    \item{\code{group}}{a factor with levels \code{IntE} \code{NE} \code{RE} \code{WE}}
  }
}
\examples{
data(groupinfo_ext)
## maybe str(groupinfo_ext) ; plot(groupinfo_ext) ...
}
\keyword{datasets}
