\name{transcript_details}
\alias{transcript_details}
\docType{data}
\title{
details of GENCODE(v 17) TSSs
}
\description{
details of GENCODE(v 17) TSSs
}
\usage{data("transcript_details")}
\format{
A data frame with 194871 observations on the following 8 variables.
}

\examples{
data(transcript_details)
## maybe str(transcript_details) ; plot(transcript_details) ...
}
\keyword{datasets}
